Move Audio DisplayObject from core to lib
